+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- method `disable_checksum()` to disable automatic CRC sum
+  appending in TX mode and checking in RX mode
 
 ## [1.0.0] - 2020-09-02
 ### Added